(f/refactor): refactored for mantainability, readability and reusability
